Add test_unzip to reproduce issue
