+++ Tests/DocUploadBundleTests/DocUploadBundleTests.swift
@@ -100,4 +100,14 @@ final class DocUploadBundleTests: XCTestCase {
                        .init(bucket: "spi-prod-docs", path: "owner/name/feature-2.0.0"))
     }
 
+    func test_unzip() async throws {
+        // https://github.com/SwiftPackageIndex/SwiftPackageIndex-Server/issues/3069
+        try await withTempDir { tempDir in
+            let url = fixtureUrl(for: "prod-apple-swift-metrics-main-e6a00d36.zip")
+            XCTAssertNoThrow(
+                try DocUploadBundle.unzip(bundle: url.path(), outputPath: tempDir)
+            )
+        }
+    }
+
 }
